Switch to mocha and playwright test
