+++ .mocharc.cjs
@@ -1,5 +1,8 @@
 module.exports = {
     extension: ['ts'],
-    spec: ['tests/**/*.node.test.ts'],
+    spec: [
+        'tests/**/*.node.test.ts',
+        'src/**/*.node.test.ts',
+    ],
     loader: 'ts-node/esm'
 }
