+++ package.json
@@ -53,10 +53,12 @@
     "build": "tsc && yarn run build:minified",
     "build:minified": "node scripts/build-minified.js",
     "start": "tsc -w",
-    "test": "TS_NODE_PROJECT=tests/tsconfig.json mocha -r ts-node/register tests/**/*.node.test.ts",
+    "test": "yarn test:node && yarn test:browser",
+    "test:node": "mocha --timeout 120000",
+    "test:browser": "playwright test",
     "test:prod": "yarn run build && yarn run lint && yarn run test:gh-action --no-cache",
     "test:types": "cp -RT tests/types/ lib/ && yarn run tsd",
-    "test:unit": "jest --watch --testPathPattern=src/",
+    "test:unit": "mocha --watch --testPathPattern=src/",
     "test:gh-action": "TEST_ENV=gh-action yarn test",
     "prepare": "yarn build && yarn docs",
     "publish-dry": "npm publish --dry-run",
