+++ src/fs/protocol/private/mmpt.node.test.ts
@@ -1,3 +1,4 @@
+import expect from "expect"
 import crypto from "crypto"
 import { IPFS } from "ipfs-core"
 
@@ -16,15 +17,17 @@ function encode(str: string): Uint8Array {
 
 let ipfs: IPFS | null = null
 
-beforeAll(async () => {
+before(async function () {
+  this.timeout(120000)
   ipfs = await createInMemoryIPFS()
   ipfsConfig.set(ipfs)
 })
 
-afterAll(async () => {
+after(async () => {
   if (ipfs == null) return
   await ipfs.stop()
 })
+
 
 /*
 Generates lots of entries for insertion into the MMPT.
@@ -35,6 +38,8 @@ This returns an array of key-values sorted by the key,
 so that key collisions are more likely to be tested.
 */
 async function generateExampleEntries(amount: number): Promise<{ name: string; cid: string }[]> {
+  if (ipfs == null) throw new Error("IPFS not loaded")
+
   const entries: { name: string; cid: string }[] = []
 
   for (const i of Array(amount).keys()) {
@@ -51,7 +56,9 @@ async function generateExampleEntries(amount: number): Promise<{ name: string; c
 
 
 
-describe("the mmpt", () => {
+describe("the mmpt", function () {
+  this.timeout(300000)
+
   it("can handle concurrent adds", async () => {
     const mmpt = MMPT.create()
 
@@ -79,7 +86,7 @@ describe("the mmpt", () => {
     const entries = await generateExampleEntries(amount)
 
     const slice_size = 5
-    let soFar = []
+    let soFar: { name: string, cid: string }[] = []
     let missing = []
 
     for (let i = 0; i < entries.length; i += slice_size) {
