+++ src/lobby/username.node.test.ts
@@ -1,3 +1,4 @@
+import expect from "expect"
 import { isUsernameValid } from "./username.js"
 
 describe("isUsernameValid", () => {
