+++ tests/fs/api.private.node.test.ts
@@ -16,7 +16,7 @@ import { privateFileContent as fileContent, privateDecode as decode } from '../h
 let ipfs: IPFS | null = null
 
 before(async function () {
-  this.timeout(10000)
+  this.timeout(120000)
   ipfs = await createInMemoryIPFS()
   ipfsConfig.set(ipfs)
 })
@@ -28,8 +28,9 @@ after(async () => {
 
 fc.configureGlobal(process.env.TEST_ENV === 'gh-action' ? { numRuns: 50 } : { numRuns: 10 })
 
-describe('the filesystem api', function () {
-  this.timeout(120000)
+describe('the private filesystem api', function () {
+  this.timeout(300000)
+
   it('writes files', async () => {
     const fs = await emptyFilesystem()
